// dcim_frontend/src/app/menu.enum.ts
export enum Menu {
  Organization = 'Organization',
  Rack_Management = 'ServerFarm',
  Device_Management = 'StockRoom'
}

export enum SubMenu {
  Locations = "Locations",
  Buildings = "Buildings",
  Racks = "Racks",
  Devices= "Devices",
  DeviceTypes = 'DeviceTypes',
  Manufacturers = 'Make',
  Models = 'Models'
}